Package lock, tsconfig, file rename (Interface)
