+++ package.json
@@ -18,8 +18,7 @@
     "firebase": "^10.12.3",
     "next": "14.2.4",
     "react": "^18",
-    "react-dom": "^18",
-    "react-icons": "^5.3.0"
+    "react-dom": "^18"
   },
   "devDependencies": {
     "@types/node": "^20",
